Fix bug where multicharacter columns were considered out of range
The comparison to see if a column was in range was incorrect. If the
columns being compared were 'AD' and 'B', then the comparison of
'B' < 'AD' was false since 'B' is less than 'A'.

This fix adds a new function to convert multicharacter columns into a
sum of their character codes so that, when we do the comparison above,
column 'B' will be considered less than 'AD'.